snapshot for error view
